backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "backtrader>=1.9.76.123",
    "fastapi>=0.90.0",
    "matplotlib>=3.0.0",
    "pandas>=1.3.0",
    "pydantic>=1.10.0,<2.0.0",
    "python-dotenv>=1.0.0",
    "tushare>=1.2.80",
    "uvicorn[standard]>=0.20.0",
]
